Move all stuff from db/InstalledPackages into Vehicle{Repository, Functions}.
